Elevator update if statements
Reversing the order of `if moving` and `if stopped` fixes issues relating movement and status: The elevator should perform stop-related operations immediately after it stops in the move-related operations.

`floor_diff` should be checked/calculated in each local section of the code, as the current floor may have changed due to move() operations in other parts of the code.